Moving demo scripts inside a src folder
