+++ demos/src/demo-legend.js
@@ -3,22 +3,17 @@
 const d3Selection = require('d3-selection');
 const PubSub = require('pubsub-js');
 
-const donut = require('./../src/charts/donut');
-const legend = require('./../src/charts/legend');
+const donut = require('./../../src/charts/donut');
+const legend = require('./../../src/charts/legend');
 
-const dataBuilder = require('./../test/fixtures/donutChartDataBuilder');
+const dataBuilder = require('./../../test/fixtures/donutChartDataBuilder');
 const colorSelectorHelper = require('./helpers/colorSelector');
 
 const dataset = new dataBuilder.DonutDataBuilder()
-        .withFivePlusOther()
-        .build();
-const datasetNoPercentages = new dataBuilder.DonutDataBuilder()
-        .withFivePlusOtherNoPercent()
-        .build();
+        .withFivePlusOther().build();
 const datasetWithThreeItems = new dataBuilder.DonutDataBuilder()
         .withThreeCategories()
         .build();
-let legendChart;
 let redrawCharts;
 
 require('./helpers/resizeHelper');
@@ -27,16 +22,11 @@ function createDonutChart(optionalColorSchema) {
     let legendChart = getLegendChart(dataset, optionalColorSchema),
         donutChart = donut(),
         donutContainer = d3Selection.select('.js-donut-chart-container'),
-        containerWidth = donutContainer.node() ? donutContainer.node().getBoundingClientRect().width : false;
+        containerWidth = d3Selection.select('.js-donut-chart-container').node().getBoundingClientRect().width;
 
     if (containerWidth) {
-        d3Selection.select('#button').on('click', function() {
-            donutChart.exportChart();
-        });
-
         donutChart
             .isAnimated(true)
-            .highlightSliceById(2)
             .width(containerWidth)
             .height(containerWidth)
             .externalRadius(containerWidth/2.5)
@@ -69,33 +59,9 @@ function getLegendChart(dataset, optionalColorSchema) {
         d3Selection.select('.js-legend-chart-container .britechart-legend').remove();
 
         legendChart
-            .width(containerWidth*0.8)
-            .height(200)
+            .height(containerWidth / 3)
+            .width(containerWidth / 2)
             .numberFormat('s');
-
-        if (optionalColorSchema) {
-            legendChart.colorSchema(optionalColorSchema);
-        }
-
-        legendContainer.datum(dataset).call(legendChart);
-
-        return legendChart;
-    }
-}
-
-function getInlineLegendChart(dataset, optionalColorSchema) {
-    let legendChart = legend(),
-        legendContainer = d3Selection.select('.js-inline-legend-chart-container'),
-        containerWidth = legendContainer.node() ? legendContainer.node().getBoundingClientRect().width : false;
-
-    if (containerWidth) {
-        d3Selection.select('.js-inline-legend-chart-container .britechart-legend').remove();
-
-        legendChart
-            .isHorizontal(true)
-            .width(containerWidth*0.6)
-            .markerSize(8)
-            .height(40)
 
         if (optionalColorSchema) {
             legendChart.colorSchema(optionalColorSchema);
@@ -130,49 +96,40 @@ function createSmallDonutChart() {
     }
 }
 
-function createDonutWithHighlightSliceChart() {
-    let donutChart = donut(),
-        donutContainer = d3Selection.select('.js-donut-highlight-slice-chart-container'),
-        containerWidth = donutContainer.node() ? donutContainer.node().getBoundingClientRect().width : false;
+function getInlineLegendChart(dataset, optionalColorSchema) {
+    let legendChart = legend(),
+        legendContainer = d3Selection.select('.js-inline-legend-chart-container'),
+        containerWidth = legendContainer.node() ? legendContainer.node().getBoundingClientRect().width : false;
 
     if (containerWidth) {
-        donutChart
-            .highlightSliceById(11)
-            .hasFixedHighlightedSlice(true)
-            .width(containerWidth)
-            .height(containerWidth/1.8)
-            .externalRadius(containerWidth/5)
-            .internalRadius(containerWidth/10);
+        d3Selection.select('.js-inline-legend-chart-container .britechart-legend').remove();
 
-        donutContainer.datum(datasetWithThreeItems).call(donutChart);
-    }
-}
+        legendChart
+            .isHorizontal(true)
+            .width(containerWidth*0.6)
+            .markerSize(8)
+            .height(40)
 
-function createLoadingState() {
-    let donutChart = donut(),
-        donutContainer = d3Selection.select('.js-loading-container'),
-        containerWidth = donutContainer.node() ? donutContainer.node().getBoundingClientRect().width : false,
-        dataset = null;
+        if (optionalColorSchema) {
+            legendChart.colorSchema(optionalColorSchema);
+        }
 
-    if (containerWidth) {
-        donutContainer.html(donutChart.loadingState());
+        legendContainer.datum(dataset).call(legendChart);
+
+        return legendChart;
     }
 }
 
 // Show charts if container available
-if (d3Selection.select('.js-donut-chart-container').node()) {
+if (d3Selection.select('.js-legend-chart-container').node()) {
     createDonutChart();
     createSmallDonutChart();
-    createDonutWithHighlightSliceChart();
-    createLoadingState();
 
     redrawCharts = function(){
         d3Selection.selectAll('.donut-chart').remove();
 
         createDonutChart();
         createSmallDonutChart();
-        createDonutWithHighlightSliceChart();
-        createLoadingState();
     };
 
     // Redraw charts on window resize
@@ -181,3 +138,4 @@ if (d3Selection.select('.js-donut-chart-container').node()) {
     // Color schema selector
     colorSelectorHelper.createColorSelector('.js-color-selector-container', '.donut-chart', createDonutChart);
 }
+
