+++ demos/src/demo-stacked-area.js
@@ -3,11 +3,11 @@
 const d3Selection = require('d3-selection');
 const PubSub = require('pubsub-js');
 
-const stackedAreaChart = require('./../src/charts/stacked-area');
-const colors = require('./../src/charts/helpers/color');
-const tooltip = require('./../src/charts/tooltip');
+const stackedAreaChart = require('./../../src/charts/stacked-area');
+const colors = require('./../../src/charts/helpers/color');
+const tooltip = require('./../../src/charts/tooltip');
 
-const stackedDataBuilder = require('./../test/fixtures/stackedAreaDataBuilder');
+const stackedDataBuilder = require('./../../test/fixtures/stackedAreaDataBuilder');
 const colorSelectorHelper = require('./helpers/colorSelector');
 let redrawCharts;
