+++ demos/src/helpers/resizeHelper.js
