+++ webpack.constants.js
@@ -20,18 +20,18 @@ exports.CHARTS = {
 };
 
 exports.DEMOS = {
-    'demo-line': './demos/demo-line.js',
-    'demo-stacked-area': './demos/demo-stacked-area.js',
-    'demo-bar': './demos/demo-bar.js',
-    'demo-grouped-bar': './demos/demo-grouped-bar.js',
-    'demo-stacked-bar': './demos/demo-stacked-bar.js',
-    'demo-donut': './demos/demo-donut.js',
-    'demo-scatter-plot': './demos/demo-scatter-plot.js',
-    'demo-sparkline': './demos/demo-sparkline.js',
-    'demo-step': './demos/demo-step.js',
-    'demo-brush': './demos/demo-brush.js',
-    'demo-kitchen-sink': './demos/demo-kitchen-sink.js',
-    'demo-legend': './demos/demo-legend.js'
+    'demo-line': './demos/src/demo-line.js',
+    'demo-stacked-area': './demos/src/demo-stacked-area.js',
+    'demo-bar': './demos/src/demo-bar.js',
+    'demo-grouped-bar': './demos/src/demo-grouped-bar.js',
+    'demo-stacked-bar': './demos/src/demo-stacked-bar.js',
+    'demo-donut': './demos/src/demo-donut.js',
+    'demo-scatter-plot': './demos/src/demo-scatter-plot.js',
+    'demo-sparkline': './demos/src/demo-sparkline.js',
+    'demo-step': './demos/src/demo-step.js',
+    'demo-brush': './demos/src/demo-brush.js',
+    'demo-kitchen-sink': './demos/src/demo-kitchen-sink.js',
+    'demo-legend': './demos/src/demo-legend.js'
 };
 
 exports.PATHS = {
